% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_co2.R
\name{add_co2}
\alias{add_co2}
\alias{add_co2.tilt_profile}
\title{Add a the jittered range of CO2 values}
\usage{
add_co2(data, co2, ...)

\method{add_co2}{tilt_profile}(
  data,
  co2,
  ...,
  jitter_amount = option_jitter_amount(),
  output_min_max = option_output_min_max(),
  output_co2_footprint = option_output_co2_footprint()
)
}
\arguments{
\item{...}{Unused but necessary for compatibility across methods.}

\item{jitter_amount}{Numeric. Controls the amount of noise. Passed to
\code{amount} in \code{\link[tiltIndicator:jitter_range]{tiltIndicator::jitter_range()}}.}

\item{output_min_max}{Logical. Output the columns \code{min} and \code{max}?}

\item{output_co2_footprint}{Logical. Output the columns \code{co2_footprint} (at
product and company levels) and \code{co2_avg} (at company level only)?}
}
\value{
An object of the same class as \code{data}.
}
\description{
Add a the jittered range of CO2 values
}
\examples{
co2 <- readr::read_csv(tiltToyData::toy_emissions_profile_products_ecoinvent())

toy_profile_emissions_impl_output() |>
  add_co2(co2)
}
\seealso{
Other composable friends: 
\code{\link{profile_emissions_impl}()}
}
\concept{composable friends}
\keyword{internal}
