% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_case_worst_case_emission_profile.R
\name{best_case_worst_case_emission_profile}
\alias{best_case_worst_case_emission_profile}
\title{Calculates best case and worst case for emission profile at product level}
\usage{
best_case_worst_case_emission_profile(data)
}
\arguments{
\item{data}{Dataframe. Emissions profile product level output}
}
\value{
A dataframe
}
\description{
Calculates best case and worst case for emission profile at product level
}
\examples{
library(tiltToyData)
library(readr)
restore <- options(readr.show_col_types = FALSE)

companies <- read_csv(toy_emissions_profile_any_companies())
co2 <- read_csv(toy_emissions_profile_products_ecoinvent())
europages_companies <- read_csv(toy_europages_companies())
ecoinvent_activities <- read_csv(toy_ecoinvent_activities())
ecoinvent_europages <- read_csv(toy_ecoinvent_europages())
isic_name <- read_csv(toy_isic_name())
profile_emissions_at_product_level <- profile_emissions_impl(
  companies,
  co2,
  europages_companies = europages_companies,
  ecoinvent_activities = ecoinvent_activities,
  ecoinvent_europages = ecoinvent_europages,
  isic = isic_name
) |>
  unnest_product()

result <- best_case_worst_case_emission_profile(profile_emissions_at_product_level)
result

# Cleanup
options(restore)
}
\keyword{internal}
