% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_case_worst_case_impl.R
\name{best_case_worst_case_impl}
\alias{best_case_worst_case_impl}
\title{Calculate preliminary best case worst case results}
\usage{
best_case_worst_case_impl(data, col_group_by, col_risk, col_rank)
}
\arguments{
\item{data}{Dataframe.}

\item{col_group_by}{Dataframe column.}

\item{col_risk}{Dataframe column.}

\item{col_rank}{Dataframe column.}
}
\description{
Calculate preliminary best case worst case results
}
\keyword{internal}
