% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_case_worst_case_sector_profile.R
\name{best_case_worst_case_sector_profile}
\alias{best_case_worst_case_sector_profile}
\title{Calculates best case and worst case for sector profile at product level}
\usage{
best_case_worst_case_sector_profile(data)
}
\arguments{
\item{data}{Dataframe. Sector profile nested output}
}
\value{
A dataframe
}
\description{
Calculates best case and worst case for sector profile at product level
}
\examples{
out <- toy_profile_sector_impl_output()

out |> best_case_worst_case_sector_profile()
}
\keyword{internal}
