% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{companies}
\alias{companies}
\title{Europages companies}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 228 rows and 11 columns.
}
\usage{
companies
}
\description{
Europages companies
}
\examples{
companies
}
\author{
Kalash Singhal.
}
\keyword{datasets}
