% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composable_friends.R, R/profile_emissions.R,
%   R/profile_emissions_upstream.R, R/profile_sector.R,
%   R/profile_sector_upstream.R
\name{composable_friends}
\alias{composable_friends}
\alias{profile_emissions_impl}
\alias{profile_emissions_upstream_impl}
\alias{profile_sector_impl}
\alias{profile_sector_upstream_impl}
\title{Composable friends}
\usage{
profile_emissions_impl(
  companies,
  co2,
  europages_companies,
  ecoinvent_activities,
  ecoinvent_europages,
  isic,
  isic_tilt = lifecycle::deprecated(),
  low_threshold = 1/3,
  high_threshold = 2/3
)

profile_emissions_upstream_impl(
  companies,
  co2,
  europages_companies,
  ecoinvent_activities,
  ecoinvent_inputs,
  ecoinvent_europages,
  isic,
  isic_tilt = lifecycle::deprecated(),
  low_threshold = 1/3,
  high_threshold = 2/3
)

profile_sector_impl(
  companies,
  scenarios,
  europages_companies,
  ecoinvent_activities,
  ecoinvent_europages,
  isic,
  isic_tilt = lifecycle::deprecated(),
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)

profile_sector_upstream_impl(
  companies,
  scenarios,
  inputs,
  europages_companies,
  ecoinvent_activities,
  ecoinvent_inputs,
  ecoinvent_europages,
  isic,
  isic_tilt = lifecycle::deprecated(),
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)
}
\description{
These functions are designed to work well together. Each does something
small. Compose them to create complex results.
}
\examples{
library(readr, warn.conflicts = FALSE)
library(tiltToyData)

restore <- options(list(
  readr.show_col_types = FALSE,
  tiltIndicatorAfter.output_co2_footprint = TRUE
))

companies <- read_csv(toy_emissions_profile_any_companies())
co2 <- read_csv(toy_emissions_profile_products_ecoinvent())
europages_companies <- read_csv(toy_europages_companies())
ecoinvent_activities <- read_csv(toy_ecoinvent_activities())
ecoinvent_europages <- read_csv(toy_ecoinvent_europages())
isic_name <- read_csv(toy_isic_name())

result <- profile_emissions_impl(
  companies,
  co2,
  europages_companies = europages_companies,
  ecoinvent_activities = ecoinvent_activities,
  ecoinvent_europages = ecoinvent_europages,
  isic = isic_name
)

result

# Cleanup
options(restore)
}
\keyword{internal}
