% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_transition_risk.R
\name{create_tr_benchmarks_tr_score}
\alias{create_tr_benchmarks_tr_score}
\title{Calulate \code{transition_risk_score} and \code{benchmark_tr_score} columns}
\usage{
create_tr_benchmarks_tr_score(data)
}
\arguments{
\item{data}{Dataframe.}
}
\description{
Calulate \code{transition_risk_score} and \code{benchmark_tr_score} columns
}
\keyword{internal}
