% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoinvent_activities}
\alias{ecoinvent_activities}
\title{Ecoinvent activities dataset}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 21238 rows and 8 columns.
}
\usage{
ecoinvent_activities
}
\description{
Ecoinvent activities dataset
}
\examples{
ecoinvent_activities
}
\author{
Kalash Singhal.
}
\keyword{datasets}
