% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoinvent_inputs}
\alias{ecoinvent_inputs}
\title{Ecoinvent inputs dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 52 rows and 3 columns.
}
\usage{
ecoinvent_inputs
}
\description{
Ecoinvent inputs dataset
}
\examples{
ecoinvent_inputs
}
\author{
Kalash Singhal.
}
\keyword{datasets}
