% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.R
\name{exclude}
\alias{exclude}
\title{Title}
\usage{
exclude(data, excluding)
}
\arguments{
\item{data}{A dataframe.}

\item{excluding}{Character vector with patterns to exclude columns and any
resulting duplicate.}
}
\description{
Title
}
\examples{
# Excludes columns along with all its duplicates
data <- tibble(x = 1, y = 1:2)
data
data |> exclude("y")

# Columns are matched as a regular expression
data <- tibble(x = 1, yz = 1:2, zy = 1)
data
data |> exclude("y")
data |> exclude("y$")
}
\seealso{
Other composable friends: 
\code{\link{jitter_co2_range}()},
\code{\link{join_to}()}
}
\concept{composable friends}
