% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ictr_product}
\alias{ictr_product}
\title{ICTR datasets}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 330 rows and 10 columns.
}
\usage{
ictr_product
}
\description{
ICTR datasets
}
\examples{
ictr_product
}
\author{
Kalash Singhal.
}
\keyword{datasets}
