% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2-inform_noise_in_co2_range.R
\name{inform_noise_in_co2_range}
\alias{inform_noise_in_co2_range}
\title{Inform the noise in the jittered CO2 range}
\usage{
inform_noise_in_co2_range(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
Called for its side effect. Returns invisible \code{data}.
}
\description{
Inform the noise in the jittered CO2 range
}
\examples{
library(tibble)

min <- 1:3
max <- 4:6
# Add 10\% noise
min_jitter <- -min * 1.1
# Add 50\% noise
max_jitter <- max * 1.5
data <- tibble(min_jitter, min, max, max_jitter)

inform_noise_in_co2_range(data)
}
\seealso{
Other composable friends: 
\code{\link{add_co2_range}()},
\code{\link{jitter_co2_range}()},
\code{\link{polish_co2_range}()},
\code{\link{score_transition_risk_and_polish}()},
\code{\link{summarize_co2_range}()}
}
\concept{composable friends}
\keyword{internal}
