% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isic}
\alias{isic}
\title{ISIC}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 182 rows and 2 columns.
}
\usage{
isic
}
\description{
ISIC
}
\examples{
isic
}
\author{
Kalash Singhal.
}
\keyword{datasets}
