% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{data}
\name{isic_name}
\alias{isic_name}
\title{Deprecated in favor of an equivalent dataset in tiltToyData}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 14 rows and 2 columns.
}
\usage{
isic_name
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
isic_name
# ->
readr::read_csv(tiltToyData::toy_isic_name())
}
\keyword{internal}
