% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isic_tilt_mapper}
\alias{isic_tilt_mapper}
\title{isic-tilt mapper}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 182 rows and 2 columns.
}
\usage{
isic_tilt_mapper
}
\description{
isic-tilt mapper
}
\examples{
isic_tilt_mapper
}
\author{
Kalash Singhal.
}
\keyword{datasets}
