% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{istr_company}
\alias{istr_company}
\title{ISTR datasets}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 36 rows and 6 columns.
}
\usage{
istr_company
}
\description{
ISTR datasets
}
\examples{
istr_company
}
\author{
Kalash Singhal.
}
\keyword{datasets}
