% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{istr_product}
\alias{istr_product}
\title{ISTR datasets}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 282 rows and 12 columns.
}
\usage{
istr_product
}
\description{
ISTR datasets
}
\examples{
istr_product
}
\author{
Kalash Singhal.
}
\keyword{datasets}
