% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2-jitter_co2_range.R
\name{jitter_co2_range}
\alias{jitter_co2_range}
\title{Expand the range of CO2 values by adding some random noise}
\usage{
jitter_co2_range(data, ...)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Arguments passed to \code{\link[tiltIndicator:jitter_range]{tiltIndicator::jitter_range()}}.}
}
\value{
A data frame.
}
\description{
This function expands a range by adding noise to the left of the minimum
values and to the right of the maximum values.
}
\examples{
library(tibble)
library(withr)

# styler: off
data <- tribble(
  ~benchmark, ~emission_profile, ~min, ~max, ~unit,
       "all",             "low",   1L,   2L,    NA,
      "unit",             "low",   3L,   4L,  "m2"
)
# styler: on

local_seed(1)
data |> jitter_co2_range()

local_seed(1)
data |> jitter_co2_range(amount = 20)

# Same
local_seed(1)
# See `?tiltIndicatorAfter_options`
local_options(tiltIndicatorAfter.set_jitter_amount = 20)
data |> jitter_co2_range(amount = set_jitter_amount())
}
\seealso{
Other composable friends: 
\code{\link{add_co2_range}()},
\code{\link{inform_noise_in_co2_range}()},
\code{\link{polish_co2_range}()},
\code{\link{score_transition_risk_and_polish}()},
\code{\link{summarize_co2_range}()}

Other functions to handle CO2 range: 
\code{\link{add_co2_range}()},
\code{\link{polish_co2_range}()},
\code{\link{summarize_co2_range}()}
}
\concept{composable friends}
\concept{functions to handle CO2 range}
\keyword{internal}
