% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{data}
\name{matches_mapper}
\alias{matches_mapper}
\title{Deprecated in favor of an equivalent dataset in tiltToyData}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 100 rows and 8 columns.
}
\usage{
matches_mapper
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
matches_mapper
# ->
readr::read_csv(tiltToyData::toy_ecoinvent_europages())
}
\keyword{internal}
