% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiltIndicatorAfter-options.R
\name{option_jitter_amount}
\alias{option_jitter_amount}
\alias{option_output_min_max}
\alias{option_output_co2_footprint}
\alias{option_verbose}
\title{Functions to get options}
\usage{
option_jitter_amount()

option_output_min_max()

option_output_co2_footprint()

option_verbose()
}
\description{
Functions to get options
}
\examples{

option_jitter_amount()

option_jitter_amount()

option_output_co2_footprint()

option_verbose()
}
\seealso{
\code{\link[=jitter_co2_range]{jitter_co2_range()}}, \link{tiltIndicatorAfter_options}
}
\keyword{internal}
