% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pctr_company}
\alias{pctr_company}
\title{PCTR datasets}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 54 rows and 4 columns.
}
\usage{
pctr_company
}
\description{
PCTR datasets
}
\examples{
pctr_company
}
\author{
Kalash Singhal.
}
\keyword{datasets}
