% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pctr_product}
\alias{pctr_product}
\title{PCTR datasets}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 30 rows and 9 columns.
}
\usage{
pctr_product
}
\description{
PCTR datasets
}
\examples{
pctr_product
}
\author{
Kalash Singhal.
}
\keyword{datasets}
