% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_best_case_worst_case.R
\name{polish_best_case_worst_case}
\alias{polish_best_case_worst_case}
\title{Polish preliminary best case worst case results}
\usage{
polish_best_case_worst_case(data)
}
\arguments{
\item{data}{Dataframe.}
}
\description{
Polish preliminary best case worst case results
}
\keyword{internal}
