% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_emissions_profile_company.R
\name{polish_emissions_profile_company}
\alias{polish_emissions_profile_company}
\title{Creates final output of Emissions Profile company level results}
\usage{
polish_emissions_profile_company(
  ep_comp,
  ep_prod,
  europages_companies,
  ecoinvent_activities,
  ecoinvent_europages,
  isic
)
}
\value{
A dataframe
}
\description{
Creates final output of Emissions Profile company level results
}
\examples{
# See examples in `?profile_emissions`
}
\keyword{internal}
