% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_emissions_profile_product.R
\name{polish_emissions_profile_product}
\alias{polish_emissions_profile_product}
\title{Creates final output of Emissions Profile product level results}
\usage{
polish_emissions_profile_product(
  ep_prod,
  europages_companies,
  ecoinvent_activities,
  ecoinvent_europages,
  isic
)
}
\value{
A dataframe
}
\description{
Creates final output of Emissions Profile product level results
}
\examples{
# See examples in `?profile_emissions`
}
\keyword{internal}
