% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_emissions_profile_upstream_product.R
\name{polish_emissions_profile_upstream_product}
\alias{polish_emissions_profile_upstream_product}
\title{Creates final output of Emissions Profile Upstream product level results}
\usage{
polish_emissions_profile_upstream_product(
  epu_prod,
  europages_companies,
  ecoinvent_activities,
  ecoinvent_europages,
  ecoinvent_inputs,
  isic
)
}
\value{
A dataframe
}
\description{
Creates final output of Emissions Profile Upstream product level results
}
\examples{
# See examples in `?profile_emissions_upstream`
}
\keyword{internal}
