% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_sector_profile_product.R
\name{polish_sector_profile_product}
\alias{polish_sector_profile_product}
\title{Creates final output of Sector Profile product level results}
\usage{
polish_sector_profile_product(
  sp_prod,
  europages_companies,
  ecoinvent_activities,
  ecoinvent_europages,
  isic
)
}
\value{
A dataframe
}
\description{
Creates final output of Sector Profile product level results
}
\examples{
# See examples in `?profile_sector_profile`
}
\keyword{internal}
