% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_sector_profile_upstream_company.R
\name{polish_sector_profile_upstream_company}
\alias{polish_sector_profile_upstream_company}
\title{Creates final output of Sector Profile Upstream company level results}
\usage{
polish_sector_profile_upstream_company(
  spu_comp,
  spu_prod,
  europages_companies,
  ecoinvent_activities,
  ecoinvent_europages,
  ecoinvent_inputs,
  isic
)
}
\value{
A dataframe
}
\description{
Creates final output of Sector Profile Upstream company level results
}
\examples{
# See examples in `?profile_sector_upstream`
}
\keyword{internal}
