% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ictr_product.R
\name{prepare_ictr_product}
\alias{prepare_ictr_product}
\title{Creates final output of ictr product level results}
\usage{
prepare_ictr_product(
  ictr_prod,
  comp,
  eco_activities,
  match_mapper,
  eco_inputs,
  isic_tilt_map
)
}
\arguments{
\item{ictr_prod}{A dataframe like \link{ictr_product}}

\item{comp}{A dataframe like \link{ep_companies}}

\item{eco_activities}{A dataframe like \link{ecoinvent_activities}}

\item{match_mapper}{A dataframe like \link{matches_mapper}}

\item{eco_inputs}{A dataframe like \link{ecoinvent_inputs}}

\item{isic_tilt_map}{A dataframe like \link{isic_tilt_mapper}}
}
\value{
A dataframe that prepares the final output of ictr_product
}
\description{
Creates final output of ictr product level results
}
\examples{
prepare_ictr_product(
  ictr_product |> head(1),
  ep_companies |> head(1),
  ecoinvent_activities |> head(1),
  matches_mapper |> head(1),
  ecoinvent_inputs |> head(1),
  isic_tilt_mapper |> head(1)
)
}
