% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_inter_pstr_product.R
\name{prepare_inter_pstr_product}
\alias{prepare_inter_pstr_product}
\title{Creates intermediate output of pstr product level results}
\usage{
prepare_inter_pstr_product(pstr_prod, comp, eco_activities, match_mapper)
}
\arguments{
\item{pstr_prod}{A dataframe like \link{pstr_product}}

\item{comp}{A dataframe like \link{ep_companies}}

\item{eco_activities}{A dataframe like \link{ecoinvent_activities}}

\item{match_mapper}{A dataframe like \link{matches_mapper}}
}
\value{
A dataframe that prepares the intermediate output of pstr_product
}
\description{
Creates intermediate output of pstr product level results
}
\examples{
matches_mapper <- matches_mapper
ecoinvent_activities <- ecoinvent_activities
pstr_product <- pstr_product
ep_companies <- ep_companies

pstr_product_inter <- prepare_inter_pstr_product(pstr_product, ep_companies, ecoinvent_activities, matches_mapper)
pstr_product_inter
}
