% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_istr_company.R
\name{prepare_istr_company}
\alias{prepare_istr_company}
\title{Creates final output of istr company level results}
\usage{
prepare_istr_company(
  istr_comp,
  istr_prod,
  comp,
  eco_activities,
  match_mapper,
  eco_inputs
)
}
\arguments{
\item{istr_comp}{A dataframe like \link{istr_company}}

\item{istr_prod}{A dataframe like \link{istr_product}}

\item{comp}{A dataframe like \link{ep_companies}}

\item{eco_activities}{A dataframe like \link{ecoinvent_activities}}

\item{match_mapper}{A dataframe like \link{matches_mapper}}

\item{eco_inputs}{A dataframe like \link{ecoinvent_inputs}}
}
\value{
A dataframe that prepares the final output of ictr_company
}
\description{
Creates final output of istr company level results
}
\examples{
matches_mapper <- matches_mapper
ecoinvent_activities <- ecoinvent_activities
istr_product <- istr_product
ep_companies <- ep_companies
istr_company <- istr_company
ecoinvent_inputs <- ecoinvent_inputs

istr_company_final <- prepare_istr_company(
  istr_company,
  istr_product,
  ep_companies,
  ecoinvent_activities,
  matches_mapper,
  ecoinvent_inputs
)
istr_company_final
}
