% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_istr_product.R
\name{prepare_istr_product}
\alias{prepare_istr_product}
\title{Creates final output of istr product level results}
\usage{
prepare_istr_product(istr_prod, comp, eco_activities, match_mapper, eco_inputs)
}
\arguments{
\item{istr_prod}{A dataframe like \link{istr_product}}

\item{comp}{A dataframe like \link{ep_companies}}

\item{eco_activities}{A dataframe like \link{ecoinvent_activities}}

\item{match_mapper}{A dataframe like \link{matches_mapper}}

\item{eco_inputs}{A dataframe like \link{ecoinvent_inputs}}
}
\value{
A dataframe that prepares the final output of istr_product
}
\description{
Creates final output of istr product level results
}
\examples{
matches_mapper <- matches_mapper
ecoinvent_activities <- ecoinvent_activities
istr_product <- istr_product
ep_companies <- ep_companies
ecoinvent_inputs <- ecoinvent_inputs

istr_product_final <- prepare_istr_product(istr_product, ep_companies, ecoinvent_activities, matches_mapper, ecoinvent_inputs)
istr_product_final
}
