% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_matches_mapper.R
\name{prepare_matches_mapper}
\alias{prepare_matches_mapper}
\title{Prepare mapper for multiple matches}
\usage{
prepare_matches_mapper(mapper, activities)
}
\arguments{
\item{mapper}{A dataframe like \link{matches_mapper}}

\item{activities}{A dataframe like \link{ecoinvent_activities}}
}
\value{
A dataframe that contains information on multiple matches
}
\description{
Prepare mapper for multiple matches
}
\examples{
matches_mapper <- matches_mapper
ecoinvent_activities <- ecoinvent_activities

output <- prepare_matches_mapper(matches_mapper, ecoinvent_activities)
output
}
