% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_pctr_company.R
\name{prepare_pctr_company}
\alias{prepare_pctr_company}
\title{Creates final output of pctr company level results}
\usage{
prepare_pctr_company(pctr_comp, pctr_prod, comp, eco_activities, match_mapper)
}
\arguments{
\item{pctr_comp}{A dataframe like \link{pctr_company}}

\item{pctr_prod}{A dataframe like \link{pctr_product}}

\item{comp}{A dataframe like \link{ep_companies}}

\item{eco_activities}{A dataframe like \link{ecoinvent_activities}}

\item{match_mapper}{A dataframe like \link{matches_mapper}}
}
\value{
A dataframe that prepares the final output of pctr_company
}
\description{
Creates final output of pctr company level results
}
\examples{
matches_mapper <- matches_mapper
ecoinvent_activities <- ecoinvent_activities
pctr_product <- pctr_product
ep_companies <- ep_companies
pctr_company <- pctr_company

pctr_company_final <- prepare_pctr_company(
  pctr_company,
  pctr_product,
  ep_companies,
  ecoinvent_activities,
  matches_mapper
)

pctr_company_final
}
