% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_pstr_product.R
\name{prepare_pstr_product}
\alias{prepare_pstr_product}
\title{Creates final output of pstr product level results}
\usage{
prepare_pstr_product(pstr_prod, comp, eco_activities, match_mapper)
}
\arguments{
\item{pstr_prod}{A dataframe like \link{pstr_product}}

\item{comp}{A dataframe like \link{ep_companies}}

\item{eco_activities}{A dataframe like \link{ecoinvent_activities}}

\item{match_mapper}{A dataframe like \link{matches_mapper}}
}
\value{
A dataframe that prepares the final output of pstr_product
}
\description{
Creates final output of pstr product level results
}
\examples{
prepare_pstr_product(
  pstr_product |> head(3),
  ep_companies |> head(3),
  ecoinvent_activities |> head(3),
  matches_mapper |> head(3)
)
}
