% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_emissions.R,
%   R/profile_emissions_upstream.R, R/profile_impl.R, R/profile_sector.R,
%   R/profile_sector_upstream.R
\name{profile_emissions_impl}
\alias{profile_emissions_impl}
\alias{profile_emissions_upstream_impl}
\alias{profile_impl}
\alias{profile_sector_impl}
\alias{profile_sector_upstream_impl}
\title{Profile implementations}
\usage{
profile_emissions_impl(
  companies,
  co2,
  europages_companies,
  ecoinvent_activities,
  ecoinvent_europages,
  isic,
  isic_tilt = lifecycle::deprecated(),
  low_threshold = 1/3,
  high_threshold = 2/3
)

profile_emissions_upstream_impl(
  companies,
  co2,
  europages_companies,
  ecoinvent_activities,
  ecoinvent_inputs,
  ecoinvent_europages,
  isic,
  isic_tilt = lifecycle::deprecated(),
  low_threshold = 1/3,
  high_threshold = 2/3
)

profile_sector_impl(
  companies,
  scenarios,
  europages_companies,
  ecoinvent_activities,
  ecoinvent_europages,
  isic,
  isic_tilt = lifecycle::deprecated(),
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)

profile_sector_upstream_impl(
  companies,
  scenarios,
  inputs,
  europages_companies,
  ecoinvent_activities,
  ecoinvent_inputs,
  ecoinvent_europages,
  isic,
  isic_tilt = lifecycle::deprecated(),
  low_threshold = ifelse(scenarios$year == 2030, 1/9, 1/3),
  high_threshold = ifelse(scenarios$year == 2030, 2/9, 2/3)
)
}
\description{
These functions are designed to work with other functions in the family of
composable friends. Compose them to create complex results.
}
\concept{composable friends}
\keyword{internal}
