% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pstr_company}
\alias{pstr_company}
\title{PSTR datasets}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 24 rows and 6 columns.
}
\usage{
pstr_company
}
\description{
PSTR datasets
}
\examples{
pstr_company
}
\author{
Kalash Singhal.
}
\keyword{datasets}
