% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_best_case_worst_case.R
\name{rename_with_prefix}
\alias{rename_with_prefix}
\title{Rename using prefix}
\usage{
rename_with_prefix(data, prefix, match = ".")
}
\arguments{
\item{data}{Dataframe.}

\item{prefix}{String.}

\item{match}{String.}
}
\description{
Rename using prefix
}
\keyword{internal}
