% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_transition_risk_and_polish.R
\name{score_transition_risk_and_polish}
\alias{score_transition_risk_and_polish}
\title{Add transition risk score and polish the output for delivery}
\usage{
score_transition_risk_and_polish(
  emissions_profile,
  sector_profile,
  pivot_wider = FALSE
)
}
\arguments{
\item{emissions_profile}{Nested data frame. The output of
\code{profile_emissions()}.}

\item{sector_profile}{Nested data frame. The output of \code{profile_sector()}.}

\item{pivot_wider}{Logical. Pivot the output at company level to a wide
format?}
}
\value{
A data frame with the column \code{companies_id}, and the nested
columns\code{product} and \code{company} holding the outputs at product and company
level.
}
\description{
Add transition risk score and polish the output for delivery
}
\examples{
library(readr, warn.conflicts = FALSE)
library(dplyr, warn.conflicts = FALSE)
library(tiltToyData, warn.conflicts = FALSE)

set.seed(123)
restore <- options(list(
  readr.show_col_types = FALSE,
  tiltIndicatorAfter.output_co2_footprint = TRUE
))

toy_emissions_profile_products_ecoinvent <- read_csv(toy_emissions_profile_products_ecoinvent())
toy_emissions_profile_any_companies <- read_csv(toy_emissions_profile_any_companies())
toy_sector_profile_any_scenarios <- read_csv(toy_sector_profile_any_scenarios())
toy_sector_profile_companies <- read_csv(toy_sector_profile_companies())
toy_europages_companies <- read_csv(toy_europages_companies())
toy_ecoinvent_activities <- read_csv(toy_ecoinvent_activities())
toy_ecoinvent_europages <- read_csv(toy_ecoinvent_europages())
toy_ecoinvent_inputs <- read_csv(toy_ecoinvent_inputs())
toy_isic_name <- read_csv(toy_isic_name())

emissions_profile <- profile_emissions_impl(
  companies = toy_emissions_profile_any_companies,
  co2 = toy_emissions_profile_products_ecoinvent,
  europages_companies = toy_europages_companies,
  ecoinvent_activities = toy_ecoinvent_activities,
  ecoinvent_europages = toy_ecoinvent_europages,
  isic = toy_isic_name
)

sector_profile <- profile_sector(
  companies = toy_sector_profile_companies,
  scenarios = toy_sector_profile_any_scenarios,
  europages_companies = toy_europages_companies,
  ecoinvent_activities = toy_ecoinvent_activities,
  ecoinvent_europages = toy_ecoinvent_europages,
  isic = toy_isic_name
)

result <- score_transition_risk_and_polish(emissions_profile, sector_profile)

result |> unnest_product()

# Most banks need company-level results in long format
long <- result |>
  unnest_company() |>
  relocate(matches("emission_"))
long

# Some banks need company-level results in wide format
wide <- score_transition_risk_and_polish(
  emissions_profile,
  sector_profile,
  pivot_wider = TRUE
) |>
  unnest_company() |>
  relocate(matches("emission_"))
wide

# Cleanup
options(restore)
}
\seealso{
Other composable friends: 
\code{\link{add_co2_range}()},
\code{\link{summarize_co2_range}()}
}
\concept{composable friends}
\keyword{internal}
