% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_profile_result.R
\name{sector_profile_result}
\alias{sector_profile_result}
\title{Example sector profile result}
\usage{
sector_profile_result()
}
\value{
A dataframe like the output of \code{\link[tiltIndicator:sector_profile]{tiltIndicator::sector_profile()}}.
}
\description{
Example sector profile result
}
\examples{
sector_profile_result()
}
\keyword{internal}
