% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jitter_co2_range.R
\name{set_jitter_amount}
\alias{set_jitter_amount}
\title{Help set the \code{amount} of jitter via \code{options()}}
\usage{
set_jitter_amount()
}
\value{
Numeric.
}
\description{
Help set the \code{amount} of jitter via \code{options()}
}
\examples{
set_jitter_amount()
}
\seealso{
\code{\link[=jitter_co2_range]{jitter_co2_range()}}, \link{tiltIndicatorAfter_options}
}
\keyword{internal}
