% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_co2_range.R
\name{summarize_co2_range}
\alias{summarize_co2_range}
\alias{jitter_co2_range}
\alias{polish_co2_range}
\title{Summarize CO2 footprint}
\usage{
summarize_co2_range(data)

jitter_co2_range(data, ...)

polish_co2_range(data)
}
\arguments{
\item{data}{The output of \verb{profile_emissions*()} at product level.}

\item{...}{Arguments passed to \link[tiltIndicator:jitter_range]{tiltIndicator::jitter_range}.}
}
\value{
A dataframe
}
\description{
Summarize CO2 footprint
}
\examples{
# styler: off
data <- tibble::tribble(
  ~benchmark, ~emission_profile, ~co2_footprint, ~unit, ~tilt_sector, ~tilt_subsector, ~isic_4digit,
       "all",             "low",             1L,  "m2",    "sector1",    "subsector1",     "'1234'",
       "all",             "low",             2L,  "m2",    "sector1",    "subsector1",     "'1234'",
       "all",            "high",             3L,  "m2",    "sector1",    "subsector1",     "'1234'",
       "all",            "high",             4L,  "m2",    "sector1",    "subsector1",     "'1234'",
      "unit",             "low",             1L,  "m2",    "sector1",    "subsector1",     "'1234'",
      "unit",             "low",             2L,  "m2",    "sector1",    "subsector1",     "'1234'",
      "unit",            "high",             3L,  "m2",    "sector1",    "subsector1",     "'1234'",
      "unit",            "high",             4L,  "m2",    "sector1",    "subsector1",     "'1234'",
)
# styler: off

data |>
  summarize_co2_range() |>
  jitter_co2_range(amount = 1) |>
  polish_co2_range()
}
\keyword{internal}
