% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiltIndicatorAfter-package.R
\docType{package}
\name{tiltIndicatorAfter-package}
\alias{tiltIndicatorAfter}
\alias{tiltIndicatorAfter-package}
\title{tiltIndicatorAfter: Create user facing outputs of tiltIndicators}
\description{
Create user facing outputs of tiltIndicators.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://2degreesinvesting.github.io/tiltIndicatorAfter/}
}

}
\author{
\strong{Maintainer}: Kalash Singhal \email{kalash@2degrees-investing.org}

Other contributors:
\itemize{
  \item Mauro Lepore (\href{https://orcid.org/0000-0002-1986-7988}{ORCID}) [contributor]
}

}
\keyword{internal}
