% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_result.R
\name{toy_output}
\alias{toy_output}
\alias{toy_emissions_profile_output}
\alias{toy_emissions_profile_upstream_output}
\alias{toy_sector_profile_output}
\alias{toy_sector_profile_upstream_output}
\title{Example datasets craeted from tiltToyData and tiltIndicator}
\usage{
toy_emissions_profile_output()

toy_emissions_profile_upstream_output()

toy_sector_profile_output()

toy_sector_profile_upstream_output()
}
\value{
A dataframe that results from applying a function from tiltIndicator
with data from tiltToyData.
}
\description{
This datasets help to develop packages that are compatible with tiltToyData
and tiltIndicator.
}
\examples{
# All datasets are momoised
system.time(toy_sector_profile_output())
system.time(toy_sector_profile_output())
}
\keyword{internal}
