% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_result.R
\name{toy_profile_emissions_impl_output}
\alias{toy_profile_emissions_impl_output}
\title{Toy output of \code{profile_emissions_impl()}}
\usage{
toy_profile_emissions_impl_output()
}
\description{
Toy output of \code{profile_emissions_impl()}
}
\examples{

toy_profile_emissions_impl_output()
}
\seealso{
Other toy outputs: 
\code{\link{toy_output}()}
}
\concept{toy outputs}
\keyword{internal}
