% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_result.R
\name{toy_result}
\alias{toy_result}
\alias{toy_emissions_profile_result}
\alias{toy_emissions_profile_upstream_result}
\alias{toy_sector_profile_result}
\alias{toy_sector_profile_upstream_result}
\title{Example datasets craeted from tiltToyData and tiltIndicator}
\usage{
toy_emissions_profile_result()

toy_emissions_profile_upstream_result()

toy_sector_profile_result()

toy_sector_profile_upstream_result()
}
\value{
A dataframe that results from applying a function from tiltIndicator
with data from tiltToyData.
}
\description{
This datasets help to develop packages that are compatible with tiltToyData
and tiltIndicator.
}
\examples{
toy_sector_profile_result()

# All datasets are momoised
system.time(toy_sector_profile_upstream_result())
system.time(toy_sector_profile_upstream_result())
}
\keyword{internal}
