% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_sector_result.R
\name{toy_sector_profile_result}
\alias{toy_sector_profile_result}
\title{Example sector profile result}
\usage{
toy_sector_profile_result()
}
\value{
A dataframe like the output of \code{\link[tiltIndicator:sector_profile]{tiltIndicator::sector_profile()}}.
}
\description{
Example sector profile result
}
\examples{
toy_sector_profile_result()
}
\keyword{internal}
