% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_risk_score.R
\name{transition_risk_score}
\alias{transition_risk_score}
\title{Transition Risk Score}
\usage{
transition_risk_score(
  emissions_profile_at_product_level,
  sector_profile_at_product_level
)
}
\arguments{
\item{emissions_profile_at_product_level}{Dataframe. Emissions profile product level output}

\item{sector_profile_at_product_level}{Dataframe. Sector profile product level output}
}
\description{
Calulate Transition Risk Score at product level and company level
}
\examples{
library(dplyr)
library(readr, warn.conflicts = FALSE)

emissions_profile_at_product_level <- example_emissions_profile_at_product_level()
sector_profile_at_product_level <- example_sector_profile_at_product_level()

result <- transition_risk_score(emissions_profile_at_product_level, sector_profile_at_product_level)

result |> unnest_product()

result |> unnest_company()
}
\seealso{
Other top-level functions: 
\code{\link{profile_emissions}()},
\code{\link{profile_sector}()}
}
\concept{top-level functions}
