% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unselect.R
\name{unselect}
\alias{unselect}
\title{Exclude columns matching a pattern}
\usage{
unselect(data, match)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See \emph{Methods}, below, for
more details.}

\item{match}{A character vector. If length > 1, the union of the
matches is taken.

For \code{starts_with()}, \code{ends_with()}, and \code{contains()} this is an exact
match. For \code{matches()} this is a regular expression, and can be a
stringr pattern.}
}
\value{
A dataframe.
}
\description{
Shortcut for \code{select(data, -matches(match))}.
}
\examples{
data <- tibble::tibble(x = 1, y = 2, yy = 3)
data |> unselect("yy")

data |> unselect("y")

data |> unselect(c("x", "yy"))
}
\seealso{
Other pipable functions: 
\code{\link{jitter_co2_range}()},
\code{\link{join_to}()}
}
\concept{pipable functions}
\keyword{internal}
