% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_co2.R
\name{add_co2}
\alias{add_co2}
\alias{add_co2.tilt_profile}
\title{Add a the jittered range of CO2 values}
\usage{
add_co2(data, co2, ...)

\method{add_co2}{tilt_profile}(
  data,
  co2,
  ...,
  jitter_amount = option_jitter_amount(),
  output_min_max = option_output_min_max(),
  output_co2_footprint = option_output_co2_footprint()
)
}
\arguments{
\item{...}{Unused but necessary for compatibility across methods.}

\item{jitter_amount}{Numeric. Controls the amount of noise. Passed to
\code{amount} in \code{\link[tiltIndicator:jitter_range]{tiltIndicator::jitter_range()}}.}

\item{output_min_max}{Logical. Output the columns \code{min} and \code{max}?}

\item{output_co2_footprint}{Logical. Output the columns \code{co2_footprint} (at
product and company levels) and \code{co2_avg} (at company level only)?}
}
\value{
An object of the same class as \code{data}.
}
\description{
Add a the jittered range of CO2 values
}
\examples{
library(readr, warn.conflicts = FALSE)
library(tiltToyData)

restore <- options(list(
  readr.show_col_types = FALSE,
  tiltIndicatorAfter.output_co2_footprint = TRUE
))

companies <- read_csv(toy_emissions_profile_any_companies())
co2 <- read_csv(toy_emissions_profile_products_ecoinvent())
europages_companies <- read_csv(toy_europages_companies())
ecoinvent_activities <- read_csv(toy_ecoinvent_activities())
ecoinvent_europages <- read_csv(toy_ecoinvent_europages())
isic_name <- read_csv(toy_isic_name())

tilt_profile <- profile_emissions_impl(
  companies,
  co2,
  europages_companies = europages_companies,
  ecoinvent_activities = ecoinvent_activities,
  ecoinvent_europages = ecoinvent_europages,
  isic = isic_name
)

tilt_profile |> add_co2(co2)

# Cleanup
options(restore)
}
\seealso{
Other composable friends: 
\code{\link{inform_noise_in_co2_range}()},
\code{\link{jitter_co2_range}()},
\code{\link{polish_co2_range}()},
\code{\link{profile_emissions_impl}()},
\code{\link{score_transition_risk_and_polish}()},
\code{\link{summarize_co2_range}()}
}
\concept{composable friends}
\keyword{internal}
