% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_co2.R
\name{add_co2}
\alias{add_co2}
\title{Add a the jittered range of CO2 values}
\usage{
add_co2(data, co2, ..., jitter_amount = option_jitter_amount())
}
\arguments{
\item{...}{Unused but necessary for compatibility across methods.}

\item{jitter_amount}{Numeric. Controls the amount of noise. Passed to
\code{amount} in \code{\link[tiltIndicator:jitter_range]{tiltIndicator::jitter_range()}}.}
}
\value{
An object of the same class as \code{data}.
}
\description{
Add a the jittered range of CO2 values
}
\examples{
library(readr)
library(tiltToyData)

co2 <- read_csv(toy_emissions_profile_products_ecoinvent())
basic_result <- toy_profile_emissions_impl_output()

basic_result |> add_co2(co2)
}
\seealso{
Other composable friends: 
\code{\link{profile_emissions_impl}()}
}
\concept{composable friends}
\concept{modules}
\keyword{internal}
