% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_co2_range.R
\name{add_co2_range}
\alias{add_co2_range}
\title{Add a the jittered range of CO2 values}
\usage{
add_co2_range(data, ...)
}
\arguments{
\item{data}{Depends on the class:
\itemize{
\item \code{data.frame}: The \code{product} data frame of a \code{tilt_profile}.
\item \code{tilt_profile}: A data frame with the column \code{companies_id}, and the list columns \code{product} and \code{company} holding the outputs at product and company level..
}}

\item{...}{Arguments passed to \link[tiltIndicator:jitter_range]{tiltIndicator::jitter_range}.}
}
\value{
A dataframe
}
\description{
\code{add_co2_range()} is a shortcut for piping \code{data} into all other
\verb{*co2_range()} functions, and then into \code{\link[tiltIndicator:join_to]{tiltIndicator::join_to()}}.
}
\examples{
library(readr, warn.conflicts = FALSE)
library(tiltToyData)

restore <- options(list(
  readr.show_col_types = FALSE,
  tiltIndicatorAfter.output_co2_footprint = TRUE
))

companies <- read_csv(toy_emissions_profile_any_companies())
co2 <- read_csv(toy_emissions_profile_products_ecoinvent())
europages_companies <- read_csv(toy_europages_companies())
ecoinvent_activities <- read_csv(toy_ecoinvent_activities())
ecoinvent_europages <- read_csv(toy_ecoinvent_europages())
isic_name <- read_csv(toy_isic_name())

out <- profile_emissions_impl(
  companies,
  co2,
  europages_companies = europages_companies,
  ecoinvent_activities = ecoinvent_activities,
  ecoinvent_europages = ecoinvent_europages,
  isic = isic_name
) |>
add_co2_range()

out |> unnest_product() |> relocate(matches("co2"))

out |> unnest_company() |> relocate(matches("co2"))

# Cleanup
options(restore)
}
\seealso{
Other composable friends: 
\code{\link{score_transition_risk_and_polish}()},
\code{\link{summarize_co2_range}()}
}
\concept{composable friends}
\keyword{internal}
