% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_case_worst_case_emission_profile.R
\name{best_case_worst_case_emission_profile}
\alias{best_case_worst_case_emission_profile}
\title{Calculates best case and worst case for emission profile at product level}
\usage{
best_case_worst_case_emission_profile(data)
}
\arguments{
\item{data}{Dataframe. Emissions profile product level output}
}
\value{
A dataframe
}
\description{
Calculates best case and worst case for emission profile at product level
}
\examples{
product <- unnest_product(toy_profile_emissions_impl_output())

product |> best_case_worst_case_emission_profile()
}
\keyword{internal}
