% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{companies}
\alias{companies}
\title{Europages companies}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 7 columns.
}
\usage{
companies
}
\description{
Europages companies
}
\examples{
companies
}
\author{
Kalash Singhal.
}
\keyword{datasets}
