% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composable_friends.R
\name{composable_friends}
\alias{composable_friends}
\title{Composable friends}
\description{
These functions are designed to work well together. Each does something
small. Compose them to create complex results.
}
\examples{
library(readr, warn.conflicts = FALSE)
library(tiltToyData)

restore <- options(list(
  readr.show_col_types = FALSE,
  tiltIndicatorAfter.output_co2_footprint = TRUE
))

companies <- read_csv(toy_emissions_profile_any_companies())
co2 <- read_csv(toy_emissions_profile_products_ecoinvent())
europages_companies <- read_csv(toy_europages_companies())
ecoinvent_activities <- read_csv(toy_ecoinvent_activities())
ecoinvent_europages <- read_csv(toy_ecoinvent_europages())
isic_name <- read_csv(toy_isic_name())

result <- profile_emissions_impl(
  companies,
  co2,
  europages_companies = europages_companies,
  ecoinvent_activities = ecoinvent_activities,
  ecoinvent_europages = ecoinvent_europages,
  isic = isic_name
)

result

# Cleanup
options(restore)
}
