% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{data}
\name{ecoinvent_activities}
\alias{ecoinvent_activities}
\title{Deprecated in favor of an equivalent dataset in tiltToyData}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 100 rows and 5 columns.
}
\usage{
ecoinvent_activities
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
ecoinvent_activities
# ->
readr::read_csv(tiltToyData::toy_ecoinvent_activities())
}
\keyword{internal}
