% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoinvent_inputs}
\alias{ecoinvent_inputs}
\title{Ecoinvent inputs dataset}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 12335 rows and 8 columns.
}
\usage{
ecoinvent_inputs
}
\description{
Ecoinvent inputs dataset
}
\examples{
ecoinvent_inputs
}
\author{
Kalash Singhal.
}
\keyword{datasets}
