% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\docType{data}
\name{ep_companies}
\alias{ep_companies}
\title{Deprecated in favor of an equivalent dataset in tiltToyData}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3 rows and 7 columns.
}
\usage{
ep_companies
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
