% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.R
\name{exclude}
\alias{exclude}
\title{Title}
\usage{
exclude(data, excluding)
}
\arguments{
\item{data}{A dataframe.}

\item{excluding}{Character vector with patterns to exclude columns and any
resulting duplicate.}
}
\description{
Title
}
