% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isic_name}
\alias{isic_name}
\title{ISIC}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 182 rows and 2 columns.
}
\usage{
isic_name
}
\description{
ISIC
}
\examples{
isic_name
}
\author{
Kalash Singhal.
}
\keyword{datasets}
