% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matches_mapper}
\alias{matches_mapper}
\title{Matcher dataset}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 100 rows and 8 columns.
}
\usage{
matches_mapper
}
\description{
Matcher dataset
}
\examples{
matches_mapper
}
\author{
Kalash Singhal.
}
\keyword{datasets}
