% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_co2_range.R
\name{polish_co2_range}
\alias{polish_co2_range}
\title{Polish the jittered range of CO2 values}
\usage{
polish_co2_range(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
An object of the same class as \code{data}.
}
\description{
Polish the jittered range of CO2 values
}
\examples{
data <- tibble(min = 1, max = 2, min_jitter = 0, max_jitter = 4)

data |> polish_co2_range()
}
\seealso{
Other composable friends: 
\code{\link{add_co2_range}()},
\code{\link{jitter_co2_range}()},
\code{\link{score_transition_risk_and_polish}()},
\code{\link{summarize_co2_range}()}
}
\concept{composable friends}
\keyword{internal}
