% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ictr_company.R
\name{prepare_ictr_company}
\alias{prepare_ictr_company}
\title{Creates final output of ictr company level results}
\usage{
prepare_ictr_company(
  ictr_comp,
  ictr_prod,
  comp,
  eco_activities,
  match_mapper,
  eco_inputs,
  isic_tilt_map
)
}
\arguments{
\item{ictr_comp}{A dataframe like \link{ictr_company}}

\item{ictr_prod}{A dataframe like \link{ictr_product}}

\item{comp}{A dataframe like \link{ep_companies}}

\item{eco_activities}{A dataframe like \link{ecoinvent_activities}}

\item{match_mapper}{A dataframe like \link{matches_mapper}}

\item{eco_inputs}{A dataframe like \link{ecoinvent_inputs}}

\item{isic_tilt_map}{A dataframe like \link{isic_tilt_mapper}}
}
\value{
A dataframe that prepares the final output of ictr_company
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
prepare_ictr_company(
  ictr_company |> head(3),
  ictr_product |> head(3),
  ep_companies |> head(3),
  ecoinvent_activities |> head(3),
  matches_mapper |> head(3),
  ecoinvent_inputs |> head(3),
  isic_tilt_mapper |> head(3)
)
}
