% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ictr_product.R
\name{prepare_ictr_product}
\alias{prepare_ictr_product}
\title{Creates final output of ictr product level results}
\usage{
prepare_ictr_product(ictr_prod, comp, eco_activities, match_mapper, eco_inputs)
}
\arguments{
\item{ictr_prod}{A dataframe like \link{ictr_product}}

\item{comp}{A dataframe like \link{companies}}

\item{eco_activities}{A dataframe like \link{ecoinvent_activities}}

\item{match_mapper}{A dataframe like \link{matches_mapper}}

\item{eco_inputs}{A dataframe like \link{ecoinvent_inputs}}
}
\value{
A dataframe that prepares the final output of ictr_product
}
\description{
Creates final output of ictr product level results
}
\examples{
matches_mapper <- matches_mapper
ecoinvent_activities <- ecoinvent_activities
ictr_product <- ictr_product
companies <- companies
ecoinvent_inputs <- ecoinvent_inputs

ictr_product <- prepare_ictr_product(ictr_product, companies, ecoinvent_activities, matches_mapper, ecoinvent_inputs)
ictr_product
}
