% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ictr_product.R
\name{prepare_ictr_product}
\alias{prepare_ictr_product}
\title{Creates final output of ictr product level results}
\usage{
prepare_ictr_product(
  ictr_prod,
  comp,
  eco_activities,
  match_mapper,
  eco_inputs,
  isic_tilt_map
)
}
\arguments{
\item{ictr_prod}{A dataframe like \link{ictr_product}}

\item{comp}{A dataframe like \link{ep_companies}}

\item{eco_activities}{A dataframe like \link{ecoinvent_activities}}

\item{match_mapper}{A dataframe like \link{matches_mapper}}

\item{eco_inputs}{A dataframe like \link{ecoinvent_inputs}}

\item{isic_tilt_map}{A dataframe like \link{isic_tilt_mapper}}
}
\value{
A dataframe that prepares the final output of ictr_product
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
# See examples in `?profile_emissions_upstream`
}
\keyword{internal}
