% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_istr_company.R
\name{prepare_istr_company}
\alias{prepare_istr_company}
\title{Creates final output of istr company level results}
\usage{
prepare_istr_company(
  istr_comp,
  istr_prod,
  comp,
  eco_activities,
  match_mapper,
  eco_inputs,
  isic_tilt_map
)
}
\arguments{
\item{comp}{A dataframe like \link{ep_companies}}

\item{eco_activities}{A dataframe like \link{ecoinvent_activities}}

\item{match_mapper}{A dataframe like \link{matches_mapper}}

\item{eco_inputs}{A dataframe like \link{ecoinvent_inputs}}

\item{isic_tilt_map}{A dataframe like \link{isic_name}}
}
\value{
A dataframe that prepares the final output of ictr_company
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
# See examples in `?profile_sector_upstream`
}
\keyword{internal}
