% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_istr_company.R
\name{prepare_istr_company}
\alias{prepare_istr_company}
\title{Creates final output of istr company level results}
\usage{
prepare_istr_company(
  istr_comp,
  istr_prod,
  comp,
  eco_activities,
  match_mapper,
  eco_inputs,
  isic_tilt_map
)
}
\arguments{
\item{istr_comp}{A dataframe like the corresponding output at company level
from tiltIndicator.}

\item{istr_prod}{A dataframe like \link{istr_product}}

\item{comp}{A dataframe like \link{ep_companies}}

\item{eco_activities}{A dataframe like \link{ecoinvent_activities}}

\item{match_mapper}{A dataframe like \link{matches_mapper}}

\item{eco_inputs}{A dataframe like \link{ecoinvent_inputs}}

\item{isic_tilt_map}{A dataframe like \link{isic_tilt_mapper}}
}
\value{
A dataframe that prepares the final output of ictr_company
}
\description{
Creates final output of istr company level results
}
