% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_istr_product.R
\name{prepare_istr_product}
\alias{prepare_istr_product}
\title{Creates final output of istr product level results}
\usage{
prepare_istr_product(
  istr_prod,
  comp,
  eco_activities,
  match_mapper,
  eco_inputs,
  isic_tilt_map
)
}
\arguments{
\item{istr_prod}{A dataframe like \link{istr_product}}

\item{comp}{A dataframe like \link{ep_companies}}

\item{eco_activities}{A dataframe like \link{ecoinvent_activities}}

\item{match_mapper}{A dataframe like \link{matches_mapper}}

\item{eco_inputs}{A dataframe like \link{ecoinvent_inputs}}

\item{isic_tilt_map}{A dataframe like \link{isic_tilt_mapper}}
}
\value{
A dataframe that prepares the final output of istr_product
}
\description{
Creates final output of istr product level results
}
\examples{
product <- unnest_product(toy_sector_profile_upstream_output())
prepare_istr_product(
  product |> head(3),
  ep_companies |> head(3),
  ecoinvent_activities |> head(3),
  matches_mapper |> head(3),
  ecoinvent_inputs |> head(3),
  isic_tilt_mapper |> head(3)
)
}
