% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_pctr_company.R
\name{prepare_pctr_company}
\alias{prepare_pctr_company}
\title{Creates final output of pctr company level results}
\usage{
prepare_pctr_company(
  pctr_comp,
  pctr_prod,
  comp,
  eco_activities,
  match_mapper,
  isic_tilt_map
)
}
\arguments{
\item{pctr_comp}{A dataframe like \link{pctr_company}}

\item{pctr_prod}{A dataframe like \link{pctr_product}}

\item{comp}{A dataframe like \link{ep_companies}}

\item{eco_activities}{A dataframe like \link{ecoinvent_activities}}

\item{match_mapper}{A dataframe like \link{matches_mapper}}

\item{isic_tilt_map}{A dataframe like \link{isic_tilt_mapper}}
}
\value{
A dataframe that prepares the final output of pctr_company
}
\description{
Creates final output of pctr company level results
}
\examples{
prepare_pctr_company(
  pctr_company |> head(1),
  pctr_product |> head(1),
  ep_companies |> head(1),
  ecoinvent_activities |> head(1),
  matches_mapper |> head(1),
  isic_tilt_mapper |> head(1)
)
}
