% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_pctr_product.R
\name{prepare_pctr_product}
\alias{prepare_pctr_product}
\title{Creates final output of pctr product level results}
\usage{
prepare_pctr_product(
  pctr_prod,
  comp,
  eco_activities,
  match_mapper,
  isic_tilt_map
)
}
\arguments{
\item{pctr_prod}{A dataframe like \link{pctr_product}}

\item{comp}{A dataframe like \link{ep_companies}}

\item{eco_activities}{A dataframe like \link{ecoinvent_activities}}

\item{match_mapper}{A dataframe like \link{matches_mapper}}

\item{isic_tilt_map}{A dataframe like \link{isic_tilt_mapper}}
}
\value{
A dataframe that prepares the final output of pctr_product
}
\description{
Creates final output of pctr product level results
}
\examples{
matches_mapper <- matches_mapper
ecoinvent_activities <- ecoinvent_activities
pctr_product <- pctr_product
ep_companies <- ep_companies
isic_tilt_mapper <- isic_tilt_mapper

pctr_product_final <- prepare_pctr_product(
  pctr_product,
  ep_companies,
  ecoinvent_activities,
  matches_mapper,
  isic_tilt_mapper
)
pctr_product_final
}
