% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_pctr_product.R
\name{prepare_pctr_product}
\alias{prepare_pctr_product}
\title{Creates final output of pctr product level results}
\usage{
prepare_pctr_product(
  pctr_prod,
  comp,
  eco_activities,
  match_mapper,
  isic_tilt_map
)
}
\arguments{
\item{comp}{A dataframe like \link{ep_companies}}

\item{eco_activities}{A dataframe like \link{ecoinvent_activities}}

\item{match_mapper}{A dataframe like \link{matches_mapper}}

\item{isic_tilt_map}{A dataframe like \link{isic_tilt_mapper}}
}
\value{
A dataframe that prepares the final output of pctr_product
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
# See examples in `?profile_emissions`
}
\keyword{internal}
