% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_pstr_company.R
\name{prepare_pstr_company}
\alias{prepare_pstr_company}
\title{Creates final output of pstr company level results}
\usage{
prepare_pstr_company(
  pstr_comp,
  pstr_prod,
  comp,
  eco_activities,
  match_mapper,
  isic_tilt_map
)
}
\arguments{
\item{comp}{A dataframe like \code{\link[tiltToyData:toy_files]{tiltToyData::toy_europages_companies()}}}

\item{eco_activities}{A dataframe like \code{\link[tiltToyData:toy_files]{tiltToyData::toy_ecoinvent_activities()}}}

\item{match_mapper}{A dataframe like \code{\link[tiltToyData:toy_files]{tiltToyData::toy_ecoinvent_europages()}}}

\item{isic_tilt_map}{A dataframe like \code{\link[tiltToyData:toy_files]{tiltToyData::toy_isic_name()}}}
}
\value{
A dataframe that prepares the final output of pstr_company
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
# See examples in `?profile_sector_profile`
}
\keyword{internal}
