% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_pstr_company.R
\name{prepare_pstr_company}
\alias{prepare_pstr_company}
\title{Creates final output of pstr company level results}
\usage{
prepare_pstr_company(pstr_comp, pstr_prod, comp, eco_activities, match_mapper)
}
\arguments{
\item{pstr_comp}{A dataframe like the corresponding output at company level
from tiltIndicator.}

\item{pstr_prod}{A dataframe like \link{pstr_product}}

\item{comp}{A dataframe like \link{ep_companies}}

\item{eco_activities}{A dataframe like \link{ecoinvent_activities}}

\item{match_mapper}{A dataframe like \link{matches_mapper}}
}
\value{
A dataframe that prepares the final output of pstr_company
}
\description{
Creates final output of pstr company level results
}
\examples{
library(tiltIndicator)

prepare_pstr_company(
  unnest_company(toy_sector_profile_result()) |> head(1),
  unnest_product(toy_sector_profile_result()) |> head(1),
  ep_companies |> head(1),
  ecoinvent_activities |> head(1),
  matches_mapper |> head(1)
)
}
