% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_pstr_company.R
\name{prepare_pstr_company}
\alias{prepare_pstr_company}
\title{Creates final output of pstr company level results}
\usage{
prepare_pstr_company(pstr_comp, pstr_prod, comp, eco_activities, match_mapper)
}
\arguments{
\item{pstr_comp}{A dataframe like the corresponding output at company level
from tiltIndicator.}

\item{pstr_prod}{A dataframe like \link{pstr_product}}

\item{comp}{A dataframe like \link{ep_companies}}

\item{eco_activities}{A dataframe like \link{ecoinvent_activities}}

\item{match_mapper}{A dataframe like \link{matches_mapper}}
}
\value{
A dataframe that prepares the final output of pstr_company
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
library(tiltIndicator)

company <- unnest_company(toy_sector_profile_upstream_output())
product <- unnest_product(toy_sector_profile_upstream_output())

prepare_pstr_company(
  company |> head(3),
  product |> head(3),
  ep_companies |> head(3),
  ecoinvent_activities |> head(3),
  matches_mapper |> head(3)
)
}
