% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_pstr_company.R
\name{prepare_pstr_company}
\alias{prepare_pstr_company}
\title{Creates final output of pstr company level results}
\usage{
prepare_pstr_company(pstr_comp, pstr_prod, comp, eco_activities, match_mapper)
}
\arguments{
\item{pstr_comp}{A dataframe like the corresponding output at company level
from tiltIndicator.}

\item{pstr_prod}{A dataframe like \link{pstr_product}}

\item{comp}{A dataframe like \link{ep_companies}}

\item{eco_activities}{A dataframe like \link{ecoinvent_activities}}

\item{match_mapper}{A dataframe like \link{matches_mapper}}
}
\value{
A dataframe that prepares the final output of pstr_company
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
See examples in `?profile_sector`
}
\keyword{internal}
