% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co2-summarize_co2_range.R
\name{summarize_co2_range}
\alias{summarize_co2_range}
\title{Summarize the range of CO2 values}
\usage{
summarize_co2_range(data)
}
\arguments{
\item{data}{Depends on the class:
\itemize{
\item \code{data.frame}: The \code{product} data frame of a \code{tilt_profile}.
\item \code{tilt_profile}: A data frame with the column \code{companies_id}, and the list columns \code{product} and \code{company} holding the outputs at product and company level..
}}
}
\value{
A data frame.
}
\description{
Summarize the range of CO2 values
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(tidyr, warn.conflicts = FALSE)

x <- tidyr::expand_grid(
  benchmark = c("all", "unit", "tilt_sector", "unit_tilt_sector"),
  emission_profile = c("low", "medium", "high"),
  unit = c("m2", "kg"),
  tilt_sector = c("sector1", "sector2"),
  tilt_subsector = c("subsector1", "subsector2"),
)
y <- tibble(
  emission_profile = c("low", "medium", "high"),
  isic_4digit = "'1234'",
  co2_footprint = 1:3,
)
data <- left_join(x, y, by = "emission_profile", relationship = "many-to-many")
data |>
  print(n = Inf)

data |>
  summarize_co2_range() |>
  print(n = Inf)
}
\seealso{
Other composable friends: 
\code{\link{add_co2_range}()},
\code{\link{inform_noise_in_co2_range}()},
\code{\link{jitter_co2_range}()},
\code{\link{polish_co2_range}()},
\code{\link{score_transition_risk_and_polish}()}

Other functions to handle CO2 range: 
\code{\link{add_co2_range}()},
\code{\link{jitter_co2_range}()},
\code{\link{polish_co2_range}()}
}
\concept{composable friends}
\concept{functions to handle CO2 range}
\keyword{internal}
