% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiltIndicatorAfter-package.R
\docType{package}
\name{tiltIndicatorAfter-package}
\alias{tiltIndicatorAfter}
\alias{tiltIndicatorAfter-package}
\title{tiltIndicatorAfter: Create user facing outputs of tiltIndicators}
\description{
Create user facing outputs of tiltIndicators.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://2degreesinvesting.github.io/tiltIndicatorAfter/}
}

}
\author{
\strong{Maintainer}: Kalash Singhal \email{kalash@2degrees-investing.org} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
