% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiltIndicatorAfter-options.R
\name{tiltIndicatorAfter_options}
\alias{tiltIndicatorAfter_options}
\title{tiltIndicatorAfter options}
\description{
Some behaviour can be controlled via an option. Typically these options are
used by developers or analysts who test the code or data before it's ready
for public consumption.
\itemize{
\item \code{tiltIndicatorAfter.co2_jitter_amount}: Controls the amount of random noise
in the \verb{co2*} columns.
\item \code{tiltIndicatorAfter.co2_keep_licensed_min_max}: Keeps the licensed \code{min}
and \code{max} columns that yield the noisy \verb{co2*} columns.
}
}
\examples{
library(readr, warn.conflicts = FALSE)
library(dplyr, warn.conflicts = FALSE)
library(withr)
library(tiltToyData)

local_options(readr.show_col_types = FALSE)

companies <- read_csv(toy_emissions_profile_any_companies())
products <- read_csv(toy_emissions_profile_products_ecoinvent())
europages_companies <- read_csv(toy_europages_companies())
ecoinvent_activities <- read_csv(toy_ecoinvent_activities())
ecoinvent_europages <- read_csv(toy_ecoinvent_europages())
isic_name <- read_csv(toy_isic_name())

amount <- 0.1
local_seed(1)
local_options(list(tiltIndicatorAfter.co2_jitter_amount = amount))
local_options(list(tiltIndicatorAfter.co2_keep_licensed_min_max = TRUE))

co2_cols <- profile_emissions(
  companies,
  products,
  europages_companies = europages_companies,
  ecoinvent_activities = ecoinvent_activities,
  ecoinvent_europages = ecoinvent_europages,
  isic = isic_name
) |>
  unnest_product() |>
  select(matches(c("min", "max", "co2")))

getOption("tiltIndicatorAfter.co2_jitter_amount")
mean(tiltIndicator::percent_noise(co2_cols$min, co2_cols$co2e_lower))
mean(tiltIndicator::percent_noise(co2_cols$max, co2_cols$co2e_upper))



# Compare
amount <- 1
local_seed(1)
local_options(list(tiltIndicatorAfter.co2_jitter_amount = amount))
local_options(list(tiltIndicatorAfter.co2_keep_licensed_min_max = TRUE))

co2_cols <- profile_emissions(
  companies,
  products,
  europages_companies = europages_companies,
  ecoinvent_activities = ecoinvent_activities,
  ecoinvent_europages = ecoinvent_europages,
  isic = isic_name
) |>
  unnest_product() |>
  select(matches(c("min", "max", "co2")))

getOption("tiltIndicatorAfter.co2_jitter_amount")
mean(tiltIndicator::percent_noise(co2_cols$min, co2_cols$co2e_lower))
mean(tiltIndicator::percent_noise(co2_cols$max, co2_cols$co2e_upper))
}
\keyword{internal}
