% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_risk_score.R
\name{transition_risk_score}
\alias{transition_risk_score}
\title{Calculate the Transition Risk Score at product and company levels
.}
\usage{
transition_risk_score(emissions_at_product_level, sector_at_product_level)
}
\arguments{
\item{emissions_at_product_level}{Emissions profile at product level result.}

\item{sector_at_product_level}{Sector profile at product level result.}
}
\description{
Calculate the Transition Risk Score at product and company levels
.
}
